Implemented Changes requested on PR#496
