+++ modules/admin-ui-picard/app/src/reducers/eventDeatilsReducers.js
@@ -73,8 +73,10 @@ import {
     SAVE_EVENT_SCHEDULING_SUCCESS,
     SAVE_EVENT_SCHEDULING_FAILURE,
     LOAD_EVENT_STATISTICS_IN_PROGRESS,
+    LOAD_EVENT_STATISTICS_SUCCESS,
     LOAD_EVENT_STATISTICS_FAILURE,
-    LOAD_EVENT_STATISTICS_SUCCESS, UPDATE_EVENT_STATISTICS_SUCCESS, UPDATE_EVENT_STATISTICS_FAILURE,
+    UPDATE_EVENT_STATISTICS_SUCCESS,
+    UPDATE_EVENT_STATISTICS_FAILURE,
 } from '../actions/eventDetailsActions';
 
 // Initial state of event details in redux store
