+++ modules/admin-ui-picard/app/src/thunks/eventDetailsThunks.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import moment from "moment";
 import {
     loadEventPoliciesInProgress,
     loadEventPoliciesSuccess,
@@ -75,16 +76,13 @@ import {
     saveEventSchedulingInProgress,
     loadEventStatisticsInProgress,
     loadEventStatisticsSuccess,
-    loadEventStatisticsFailure, updateEventStatisticsSuccess, updateEventStatisticsFailure,
+    loadEventStatisticsFailure,
+    updateEventStatisticsSuccess,
+    updateEventStatisticsFailure,
 } from '../actions/eventDetailsActions';
+import {removeNotificationWizardForm} from "../actions/notificationActions";
 import {addNotification} from "./notificationThunks";
-import {
-    createPolicy,
-    getHttpHeaders,
-    transformMetadataCollection,
-    transformMetadataForUpdate
-} from "../utils/resourceUtils";
-import {NOTIFICATION_CONTEXT} from "../configs/modalConfig";
+import {fetchWorkflowDef} from "./workflowThunks";
 import {
     getBaseWorkflow,
     getCaptureAgents,
@@ -93,15 +91,23 @@ import {
     getSchedulingSource,
     getWorkflow,
     getWorkflowDefinitions,
-    getWorkflows, getStatistics
+    getWorkflows,
+    getStatistics
 } from "../selectors/eventDetailsSelectors";
-import {fetchWorkflowDef} from "./workflowThunks";
 import {getWorkflowDef} from "../selectors/workflowSelectors";
-import {logger} from "../utils/logger";
-import {removeNotificationWizardForm} from "../actions/notificationActions";
+import {NOTIFICATION_CONTEXT} from "../configs/modalConfig";
+import {
+    createPolicy,
+    getHttpHeaders,
+    transformMetadataCollection,
+    transformMetadataForUpdate
+} from "../utils/resourceUtils";
+import {
+    createChartOptions,
+    createDownloadUrl
+} from "../utils/statisticsUtils";
 import {calculateDuration} from "../utils/dateUtils";
-import moment from "moment";
-import {getCurrentLanguageInformation} from "../utils/utils";
+import {logger} from "../utils/logger";
 
 
 // thunks for metadata
@@ -1061,107 +1067,6 @@ export const fetchEventPublications = eventId => async dispatch => {
 
 // thunks for statistics
 
-/* creates callback function for formatting the labels of the xAxis in a statistics diagram */
-const createXAxisTickCallback = (timeMode, dataResolution, language) => {
-
-    return (value, index, ticks) => {
-        let formatString = 'L';
-        if (timeMode === 'year') {
-            formatString = 'MMMM';
-        } else if (timeMode === 'month') {
-            formatString = 'dddd, Do';
-        } else {
-            if (dataResolution === 'yearly') {
-                formatString = 'YYYY';
-            } else if (dataResolution === 'monthly') {
-                    formatString = 'MMMM';
-            } else if (dataResolution === 'daily') {
-                formatString = 'MMMM Do, YYYY';
-            } else if (dataResolution === 'hourly') {
-                formatString = 'LLL';
-            }
-        }
-
-        return moment(value).locale(language).format(formatString);
-    }
-}
-
-/* creates callback function for the displayed label when hovering over a data point in a statistics diagram */
-const createTooltipCallback = (chooseMode, dataResolution, language) => {
-    return (tooltipItem) => {
-        const date = tooltipItem.label;
-
-        let formatString;
-        if (chooseMode === 'year') {
-            formatString = 'MMMM YYYY';
-        } else if (chooseMode === 'month') {
-            formatString = 'dddd, MMMM Do, YYYY';
-        } else {
-            if (dataResolution === 'yearly') {
-                formatString = 'YYYY';
-            } else if (dataResolution === 'monthly') {
-                    formatString = 'MMMM YYYY';
-            } else if (dataResolution === 'daily') {
-                formatString = 'dddd, MMMM Do, YYYY';
-            } else {
-                formatString = 'dddd, MMMM Do, YYYY HH:mm';
-            }
-        }
-        const finalDate = moment(date).locale(language).format(formatString);
-        return finalDate + ': ' + tooltipItem.value;
-    }
-}
-
-/* creates options for statistics chart */
-const createChartOptions = (timeMode, dataResolution) => {
-
-    // Get info about the current language and its date locale
-    const currentLanguage = getCurrentLanguageInformation().dateLocale.code;
-
-    return {
-        responsive: true,
-        legend: {
-            display: false
-        },
-        layout: {
-            padding: {
-                top: 20,
-                left: 20,
-                right: 20
-            }
-        },
-        scales: {
-            xAxes: [{
-                ticks: {
-                    callback: createXAxisTickCallback(timeMode, dataResolution, currentLanguage)
-                }
-            }],
-            y: {
-                suggestedMin: 0
-            }
-        },
-        tooltips: {
-            callbacks: {
-                label: createTooltipCallback(timeMode, dataResolution, currentLanguage)
-            }
-        }};
-}
-
-/* creates the url for downloading a csv file with current statistics */
-const createDownloadUrl = (eventId, providerId, from, to, dataResolution) => {
-    const csvUrlSearchParams = new URLSearchParams({
-        dataResolution: dataResolution,
-        providerId: providerId,
-        resourceId: eventId,
-        resourceType: 'episode',
-        from: moment(from).toJSON(),
-        to: moment(to).endOf('day').toJSON()
-    });
-
-    return '/admin-ng/statistics/export.csv?' + csvUrlSearchParams;
-
-}
-
 export const fetchStatistics = eventId => async (dispatch, getState) => {
     dispatch(loadEventStatisticsInProgress());
 
@@ -1189,7 +1094,7 @@ export const fetchStatistics = eventId => async (dispatch, getState) => {
             // iterate over statistics providers
             for(let i = 0; i < response.data.length; i++){
 
-                // currently only time series data can be displayed, for other types, add data directly, then continue
+                // currently, only time series data can be displayed, for other types, add data directly, then continue
                 if(response.data[i].providerType !== 'timeSeries'){
                     newStatistics.push({
                         ...response.data[i]
